requirements.txt
cohere==2.9.1
fastapi==0.88.0
pinecone==0.1.0
pinecone_client==2.0.13
python-dotenv==0.21.0
uvicorn==0.20.0
